hotelpilot_langchain.py
#!/usr/bin/env python
"""
HotelPilot - LangChain Version
===============================
Using LangChain for cleaner agent orchestration.
"""

import os
from typing import List, Dict, Any
from dotenv import load_dotenv
from dataclasses import dataclass

# Load environment
load_dotenv('hotelpilot/.env')

# LangChain approach (pseudo-code - install langchain first)
"""
To use this version, install:
pip install langchain langchain-google-genai
"""

try:
    from langchain.agents import AgentExecutor, create_react_agent
    from langchain.memory import ConversationBufferMemory
    from langchain.prompts import PromptTemplate
    from langchain_google_genai import ChatGoogleGenerativeAI
    from langchain.tools import Tool
    LANGCHAIN_AVAILABLE = True
except ImportError:
    LANGCHAIN_AVAILABLE = False
    print("LangChain not installed. Using simple orchestration instead.")

# Simple Chain-of-Thought Orchestrator
class SimpleChainOrchestrator:
    """Simplified Chain-of-Thought orchestrator without heavy dependencies"""

    def __init__(self):
        import google.generativeai as genai
        genai.configure(api_key=os.getenv('GOOGLE_AI_API_KEY'))
        self.model = genai.GenerativeModel('gemini-2.5-flash')
        self.agents = {}
        self.memory = []  # Simple conversation memory
        self._setup_agents()

    def _setup_agents(self):
        """Setup agent configurations"""
        self.agents = {
            "router": {
                "role": "Request Router",
                "prompt": """You are a request router. Analyze the user request and determine:
                1. Which agent(s) should handle it
                2. In what order
                3. What information to pass between them

                Agents available:
                - ops_copilot: Overall operations
                - demand: Pricing decisions
                - guest: Communications
                - housekeeping: Room operations
                - billing: Payments

                Output format:
                AGENTS: [list of agents]
                REASON: why this routing
                """
            },
            "ops_copilot": {
                "role": "Operations Supervisor",
                "prompt": """You are the Ops Copilot for Arlington Hotel.
                Think step by step:
                1. What is the user asking?
                2. What information do I need?
                3. What's my recommendation?

                Be concise and actionable."""
            },
            "demand": {
                "role": "Revenue Manager",
                "prompt": """You are the Revenue Manager.
                Always:
                1. Show current rates
                2. Calculate demand lift
                3. Apply rate caps (weekday 10%, weekend 12%)
                4. Give specific price recommendation"""
            },
            "guest": {
                "role": "Guest Relations",
                "prompt": """You manage guest communications.
                Format all messages as:
                - Subject: [topic]
                - Message: [concise content]
                - Tone: Warm and professional"""
            },
            "housekeeping": {
                "role": "Housekeeping Manager",
                "prompt": """You manage room operations.
                Always include:
                - Time estimates
                - Staff assignments
                - Priority order"""
            },
            "billing": {
                "role": "Payment Specialist",
                "prompt": """You handle payments.
                For each issue provide:
                - Recovery strategy
                - Timeline
                - Success probability"""
            }
        }

    def think(self, agent_name: str, user_input: str, context: str = "") -> str:
        """Agent thinking process with Chain-of-Thought"""
        agent = self.agents.get(agent_name, self.agents["ops_copilot"])

        thought_prompt = f"""Role: {agent['role']}

{agent['prompt']}

Context: {context if context else 'First interaction'}
User Input: {user_input}

Think through this step-by-step and provide your response:"""

        response = self.model.generate_content(thought_prompt)
        return response.text

    def route_and_process(self, user_input: str) -> Dict[str, Any]:
        """Route and process with Chain-of-Thought"""

        # Step 1: Routing decision
        routing = self.think("router", user_input)

        # Extract agents (simple parsing)
        if "demand" in routing.lower() or "pric" in user_input.lower():
            primary_agent = "demand"
        elif "guest" in routing.lower() or "message" in user_input.lower():
            primary_agent = "guest"
        elif "clean" in user_input.lower() or "room" in user_input.lower():
            primary_agent = "housekeeping"
        elif "pay" in user_input.lower() or "bill" in user_input.lower():
            primary_agent = "billing"
        else:
            primary_agent = "ops_copilot"

        # Step 2: Process with primary agent
        response = self.think(primary_agent, user_input, str(self.memory[-3:]))

        # Step 3: Store in memory
        self.memory.append({
            "user": user_input,
            "agent": primary_agent,
            "response": response
        })

        return {
            "agent": self.agents[primary_agent]["role"],
            "thinking": routing,
            "response": response,
            "memory_size": len(self.memory)
        }

    def multi_step(self, user_input: str, steps: List[str]) -> Dict[str, Any]:
        """Multi-step processing through multiple agents"""
        results = []
        context = user_input

        for step in steps:
            result = self.think(step, user_input, context)
            results.append({
                "agent": self.agents[step]["role"],
                "output": result
            })
            context += f"\n\n{self.agents[step]['role']} says: {result}"

        # Final summary by Ops Copilot
        summary_prompt = f"""Summarize these multi-agent responses into an action plan:

Original request: {user_input}

Agent responses:
{context}

Provide a clear, actionable summary:"""

        summary = self.model.generate_content(summary_prompt)

        return {
            "steps": results,
            "summary": summary.text
        }

# If LangChain is available, use it
if LANGCHAIN_AVAILABLE:
    class LangChainHotelOrchestrator:
        """Full LangChain implementation"""

        def __init__(self):
            self.llm = ChatGoogleGenerativeAI(
                model="gemini-2.5-flash",
                temperature=0.7
            )
            self.memory = ConversationBufferMemory()
            self.agents = self._create_agents()

        def _create_agents(self):
            """Create LangChain agents"""
            # This would include full LangChain agent setup
            # with tools, chains, and proper routing
            pass

        def process(self, user_input: str):
            """Process with LangChain agents"""
            # Full LangChain processing
            pass

# Main execution
def main():
    print("""
    ╔═══════════════════════════════════════╗
    ║   HotelPilot - Simple Orchestration   ║
    ║         No Complex Dependencies       ║
    ╚═══════════════════════════════════════╝
    """)

    # Use simple orchestrator
    orchestrator = SimpleChainOrchestrator()

    # Test examples
    examples = [
        "What are the rates for tomorrow?",
        "Create a welcome message for VIP guest John Smith",
        "Schedule housekeeping for 20 checkouts",
        "Handle failed payment of $500"
    ]

    for example in examples:
        print(f"\n{'='*50}")
        print(f"Query: {example}")
        print("-"*50)
        result = orchestrator.route_and_process(example)
        print(f"Agent: {result['agent']}")
        print(f"Response: {result['response'][:200]}...")

    # Interactive mode
    print("\n\nInteractive Mode ('quit' to exit)")
    print("="*50)

    while True:
        user_input = input("\n> ")
        if user_input.lower() == 'quit':
            break

        result = orchestrator.route_and_process(user_input)
        print(f"\n[{result['agent']}]:")
        print(result['response'])

if __name__ == "__main__":
    main()